--- lib/cubit/juz/juz_cubit_logic.dart
@@ -1,15 +1,0 @@
-import 'package:flutter/material.dart';
-
-class JuzCubicLogic extends StatefulWidget {
-  const JuzCubicLogic({super.key});
-
-  @override
-  State<JuzCubicLogic> createState() => _JuzCubicLogicState();
-}
-
-class _JuzCubicLogicState extends State<JuzCubicLogic> {
-  @override
-  Widget build(BuildContext context) {
-    return const Placeholder();
-  }
-}
