+++ lib/cubit/juz/juz_state.dart
